requirements.txt
aiofiles==0.5.0
aniso8601==7.0.0
async-exit-stack==1.0.1
async-generator==1.10
asyncpg==0.23.0
certifi==2021.5.30
chardet==4.0.0
click==7.1.2
databases==0.4.3
dnspython==2.1.0
email-validator==1.1.3
fastapi==0.65.2
GeoAlchemy2==0.9.1
graphene==2.1.8
graphql-core==2.3.2
graphql-relay==2.0.1
greenlet==1.1.0
h11==0.12.0
httptools==0.1.2
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==2.0.1
orjson==3.5.3
pytest==6.2.4
promise==2.3
psycopg2==2.9.1
pydantic==1.8.2
pyproj==3.1.0
python-dotenv==0.18.0
python-multipart==0.0.5
PyYAML==5.4.1
requests==2.25.1
Rx==1.6.1
Shapely==1.7.1
six==1.16.0
SQLAlchemy==1.3.24
starlette==0.14.2
typing-extensions==3.10.0.0
ujson==4.0.2
urllib3==1.26.6
uvicorn==0.13.4
uvloop==0.15.2
watchgod==0.7
websockets==8.1
